refactor: refine the breakdown of `EditorTable.ts`
- Rename `Table` to `CressTable`
- Rename `ColumnHelper` to `ColumnTools`
- Rename `CressTable::beforeValidate` and `CressTable::afterValidate` to `CressTable::setProcessStatus` and `CressTable::setResultStatus`
- Reformat files

Refs: #73
