+++ tsconfig.json
@@ -1,18 +1,13 @@
 {
-    "compilerOptions": {
-        "outDir": "./build",
-        "allowJs": true,
-        "target": "es2017",
-        "sourceMap": true,
-        "module": "commonjs",
-        "allowSyntheticDefaultImports": true,
-        "resolveJsonModule": true
-    },
-    "include": [
-        "./src/**/*.ts",
-        "./test/**/*.ts"
-    ],
-    "exclude": [
-        "node_modules"
-    ]
+  "compilerOptions": {
+    "outDir": "./build",
+    "allowJs": true,
+    "target": "es2017",
+    "sourceMap": true,
+    "module": "commonjs",
+    "allowSyntheticDefaultImports": true,
+    "resolveJsonModule": true
+  },
+  "include": ["./src/**/*.ts", "./test/**/*.ts"],
+  "exclude": ["node_modules"]
 }
